requirements.txt
# requirements.txt
# Proje  : Eğitim Teknolojileri için PyTorch Eğitimi
# Yazar  : Ugur Sirvermez — Bursa Uludag Universitesi
#
# Kurulum:
#   pip install -r requirements.txt
#
# Not: torch ve torchvision için CUDA sürümünüze uygun komutu
#   https://pytorch.org/get-started/locally/ adresinden alınız.
#   Örnek (CPU): pip install torch torchvision --index-url https://download.pytorch.org/whl/cpu

# ── Temel: PyTorch ──────────────────────────────────────────────────
torch>=2.0.0
torchvision>=0.15.0

# ── Görüntü İşleme ──────────────────────────────────────────────────
Pillow>=9.0.0

# ── Veri + Sayısal İşlemler ─────────────────────────────────────────
numpy>=1.23.0
pandas>=1.5.0

# ── Görselleştirme ───────────────────────────────────────────────────
matplotlib>=3.6.0

# ── İlerleme Çubuğu ─────────────────────────────────────────────────
tqdm>=4.64.0

# ── HTTP İstekleri (veri indirme) ────────────────────────────────────
requests>=2.28.0

# ── Model Özeti ──────────────────────────────────────────────────────
torchinfo>=1.7.0

# ── Transfer Learning Metrikleri ─────────────────────────────────────
torchmetrics>=0.11.0

# ── Sklearn Yardımcıları (karışıklık matrisi vb.) ────────────────────
scikit-learn>=1.1.0

# ── Karışıklık Matrisi Görselleştirme ────────────────────────────────
mlxtend>=0.21.0

# ── Pekiştirmeli Öğrenme (Modül 11) ─────────────────────────────────
gymnasium>=0.28.0
imageio>=2.22.0
pyvirtualdisplay>=3.0

# ── Gradio Arayüzü (Modül 09) ────────────────────────────────────────
gradio>=3.40.0

# ── Jupyter / Colab ──────────────────────────────────────────────────
ipython>=7.0.0
notebook>=6.5.0
jupyterlab>=3.5.0
ipykernel>=6.0.0
